moved custom activations to separate file
