+++ src/luxonis_train/models/backbones/micronet.py
@@ -5,24 +5,7 @@
 
 import torch
 from torch import nn, Tensor
-
-
-class HSigmoid(nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.relu = nn.ReLU6(True)
-
-    def forward(self, x: Tensor) -> Tensor:
-        return self.relu(x + 3) / 6
-
-
-class HSwish(nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.sigmoid = HSigmoid()
-
-    def forward(self, x: Tensor) -> Tensor:
-        return x * self.sigmoid(x)
+from luxonis_train.models.modules import HSigmoid, HSwish
 
 
 class ChannelShuffle(nn.Module):
