+++ src/luxonis_train/models/modules/__init__.py
@@ -1,1 +1,2 @@
 from .common import *
+from .activations import *
